Fix spotless
