+++ src/commonMain/kotlin/uk/gov/hmrc/calculator/model/bands/TaxBands.kt
@@ -18,7 +18,6 @@ package uk.gov.hmrc.calculator.model.bands
 import uk.gov.hmrc.calculator.exception.InvalidTaxYearException
 import uk.gov.hmrc.calculator.model.Country
 import uk.gov.hmrc.calculator.model.Country.SCOTLAND
-import uk.gov.hmrc.calculator.model.Country.WALES
 
 internal class TaxBands(country: Country, taxYear: Int) {
